Split compiler internals into a separate file
