+++ apps/api-extractor/src/analyzer/TypeScriptHelpers.ts
@@ -5,6 +5,7 @@
 
 import * as ts from 'typescript';
 import { TypeScriptMessageFormatter } from './TypeScriptMessageFormatter';
+import { TypeScriptInternals } from './TypeScriptInternals';
 
 export class TypeScriptHelpers {
   /**
@@ -32,10 +33,6 @@ export class TypeScriptHelpers {
     }
 
     return current;
-  }
-
-  public static getImmediateAliasedSymbol(symbol: ts.Symbol, typeChecker: ts.TypeChecker): ts.Symbol {
-    return (typeChecker as any).getImmediateAliasedSymbol(symbol); // tslint:disable-line:no-any
   }
 
   /**
@@ -70,26 +67,11 @@ export class TypeScriptHelpers {
   }
 
   /**
-   * Returns the Symbol for the provided Declaration.  This is a workaround for a missing
-   * feature of the TypeScript Compiler API.   It is the only apparent way to reach
-   * certain data structures, and seems to always work, but is not officially documented.
-   *
-   * @returns The associated Symbol.  If there is no semantic information (e.g. if the
-   * declaration is an extra semicolon somewhere), then "undefined" is returned.
-   */
-  public static tryGetSymbolForDeclaration(declaration: ts.Declaration): ts.Symbol | undefined {
-    /* tslint:disable:no-any */
-    const symbol: ts.Symbol = (declaration as any).symbol;
-    /* tslint:enable:no-any */
-    return symbol;
-  }
-
-  /**
    * Same semantics as tryGetSymbolForDeclaration(), but throws an exception if the symbol
    * cannot be found.
    */
   public static getSymbolForDeclaration(declaration: ts.Declaration): ts.Symbol {
-    const symbol: ts.Symbol | undefined = TypeScriptHelpers.tryGetSymbolForDeclaration(declaration);
+    const symbol: ts.Symbol | undefined = TypeScriptInternals.tryGetSymbolForDeclaration(declaration);
     if (!symbol) {
       throw new Error(TypeScriptMessageFormatter.formatFileAndLineNumber(declaration) + ': '
         + 'Unable to determine semantic information for this declaration');
@@ -103,46 +85,10 @@ export class TypeScriptHelpers {
 
     if (declarationWithModuleSpecifier.moduleSpecifier
       && ts.isStringLiteralLike(declarationWithModuleSpecifier.moduleSpecifier)) {
-      return TypeScriptHelpers.getTextOfIdentifierOrLiteral(declarationWithModuleSpecifier.moduleSpecifier);
+      return TypeScriptInternals.getTextOfIdentifierOrLiteral(declarationWithModuleSpecifier.moduleSpecifier);
     }
 
     return undefined;
-  }
-
-  /**
-   * Retrieves the comment ranges associated with the specified node.
-   */
-  public static getJSDocCommentRanges(node: ts.Node, text: string): ts.CommentRange[] | undefined {
-    // Compiler internal:
-    // https://github.com/Microsoft/TypeScript/blob/v2.4.2/src/compiler/utilities.ts#L616
-
-    // tslint:disable-next-line:no-any
-    return (ts as any).getJSDocCommentRanges.apply(this, arguments);
-  }
-
-  /**
-   * Retrieves the (unescaped) value of an string literal, numeric literal, or identifier.
-   */
-  public static getTextOfIdentifierOrLiteral(node: ts.Identifier | ts.StringLiteralLike | ts.NumericLiteral): string {
-    // Compiler internal:
-    // https://github.com/Microsoft/TypeScript/blob/v3.2.2/src/compiler/utilities.ts#L2721
-
-    // tslint:disable-next-line:no-any
-    return (ts as any).getTextOfIdentifierOrLiteral(node);
-  }
-
-  /**
-   * Retrieves the (cached) module resolution information for a module name that was exported from a SourceFile.
-   * The compiler populates this cache as part of analyzing the source file.
-   */
-  public static getResolvedModule(sourceFile: ts.SourceFile, moduleNameText: string): ts.ResolvedModuleFull
-    | undefined {
-
-    // Compiler internal:
-    // https://github.com/Microsoft/TypeScript/blob/v3.2.2/src/compiler/utilities.ts#L218
-
-    // tslint:disable-next-line:no-any
-    return (ts as any).getResolvedModule(sourceFile, moduleNameText);
   }
 
   /**
